LayoutTests/media/audio-garbage-collect-expected.txt

Tests that we don't garbage collect playing audio object or event listener.

According to http://www.whatwg.org/specs/web-apps/current-work/multipage/the-video-element.html,
"4.8.10.8 Playing the media resource",
"Media elements must not stop playing just because all references to them have been removed; only once a media element is in a state where no further audio could ever be played by that element may the element be garbage collected."

(see https://bugs.webkit.org/show_bug.cgi?id=66878, https://bugs.webkit.org/show_bug.cgi?id=70421, and http://crbug.com/62604 for more details).

PASS
